+++ src/js/core/scroll.js
@@ -6,13 +6,13 @@ export default function (UIkit) {
 
         props: {
             duration: Number,
-            transition: String,
+            easing: String,
             offset: Number
         },
 
         defaults: {
             duration: 1000,
-            transition: 'easeOutExpo',
+            easing: 'easeOutExpo',
             offset: 0
         },
 
@@ -20,10 +20,8 @@ export default function (UIkit) {
 
             scrollToElement(el) {
 
-                el = $(el);
-
                 // get / set parameters
-                var target = offsetTop(el) - this.offset,
+                var target = offsetTop($(el)) - this.offset,
                     docHeight = document.documentElement.offsetHeight,
                     winHeight = window.innerHeight;
 
@@ -34,7 +32,7 @@ export default function (UIkit) {
                 // animate to target, fire callback when done
                 $('html,body')
                     .stop()
-                    .animate({scrollTop: Math.round(target)}, this.duration, this.transition)
+                    .animate({scrollTop: Math.round(target)}, this.duration, this.easing)
                     .promise()
                     .then(() => this.$el.trigger('scrolled', [this]));
 
@@ -58,10 +56,8 @@ export default function (UIkit) {
 
     });
 
-    if (!$.easing.easeOutExpo) {
-        $.easing.easeOutExpo = function (x, t, b, c, d) {
-            return (t === d) ? b + c : c * (-Math.pow(2, -10 * t / d) + 1) + b;
-        };
-    }
+    $.easing.easeOutExpo = $.easing.easeOutExpo || function (x, t, b, c, d) {
+        return (t === d) ? b + c : c * (-Math.pow(2, -10 * t / d) + 1) + b;
+    };
 
 }