+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Changed
 
 - IMPORTANT: Move Less/Sass imports into one file. Use `components/_import.less`
+- IMPORTANT: Renamed 'transition' option to 'easing' in Scroll component
 - Improved flip behaviour in Drop component
 
 ### Fixed